requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
wordcloud
scikit-learn
opencv-python
Pillow
plotly
nltk
